func/llamaindex/requirements.txt
llama-index
llama-index-llms-azure-openai
chainlit
python-dotenv